+++ src/snowflake/cli/_plugins/sql/manager.py
@@ -14,22 +14,23 @@
 
 from __future__ import annotations
 
+import logging
 import sys
-from io import StringIO
-from itertools import chain
+from functools import partial
 from pathlib import Path
 from typing import Dict, Iterable, List, Tuple
 
 from click import ClickException, UsageError
-from jinja2 import UndefinedError
+from snowflake.cli._plugins.sql.reader import SQLReader
 from snowflake.cli._plugins.sql.snowsql_templating import transpile_snowsql_templates
+from snowflake.cli.api.console import cli_console
 from snowflake.cli.api.rendering.sql_templates import snowflake_sql_jinja_render
-from snowflake.cli.api.secure_path import UNLIMITED, SecurePath
 from snowflake.cli.api.sql_execution import SqlExecutionMixin, VerboseCursor
 from snowflake.connector.cursor import SnowflakeCursor
-from snowflake.connector.util_text import split_statements
 
 IsSingleStatement = bool
+
+logger = logging.getLogger(__name__)
 
 
 class SqlManager(SqlExecutionMixin):
@@ -41,57 +42,38 @@ class SqlManager(SqlExecutionMixin):
         data: Dict | None = None,
         retain_comments: bool = False,
     ) -> Tuple[IsSingleStatement, Iterable[SnowflakeCursor]]:
-        inputs = [query, files, std_in]
-        # Check if any two inputs were provided simultaneously
-        if len([i for i in inputs if i]) > 1:
-            raise UsageError(
-                "Multiple input sources specified. Please specify only one."
-            )
+        """Reads, transforms and execute statements from input.
 
-        if std_in:
-            query = sys.stdin.read()
-        if query:
-            return self._execute_single_query(
-                query=query, data=data, retain_comments=retain_comments
-            )
+        Only one input can be consumed at a time. If multiple inputs are provided, the
+        order of precedence is as follows:
+        - stdin
+        - query
+        - files
 
-        if files:
-            # Multiple files
-            results = []
-            single_statement = False
-            for file in files:
-                query_from_file = SecurePath(file).read_text(
-                    file_size_limit_mb=UNLIMITED
-                )
-                single_statement, result = self._execute_single_query(
-                    query=query_from_file, data=data, retain_comments=retain_comments
-                )
-                results.append(result)
+        When no compilation errors are detected, the sequence on queries
+        in executed and returned as tuple.
+        """
+        query = sys.stdin.read() if std_in else query
 
-            # Use single_statement if there's only one, otherwise this is multi statement result
-            single_statement = len(files) == 1 and single_statement
-            return single_statement, chain.from_iterable(results)
-
-        # At that point, no stdin, query or files were provided
-        raise UsageError("Use either query, filename or input option.")
-
-    def _execute_single_query(
-        self, query: str, data: Dict | None = None, retain_comments: bool = False
-    ) -> Tuple[IsSingleStatement, Iterable[SnowflakeCursor]]:
-        try:
-            query = transpile_snowsql_templates(query)
-            query = snowflake_sql_jinja_render(content=query, data=data)
-        except UndefinedError as err:
-            raise ClickException(f"SQL template rendering error: {err}")
-
-        statements = tuple(
-            statement
-            for statement, _ in split_statements(
-                StringIO(query), remove_comments=not retain_comments
-            )
+        stmt_reader = SQLReader(query, files, not retain_comments)
+        stmt_operator_funcs = (
+            transpile_snowsql_templates,
+            partial(snowflake_sql_jinja_render, data=data),
         )
-        single_statement = len(statements) == 1
+        errors, stmt_count, compiled_statements = stmt_reader.compile_statements(
+            stmt_operator_funcs
+        )
+        if not any((errors, stmt_count, compiled_statements)):
+            raise UsageError("Use either query, filename or input option.")
 
-        return single_statement, self._execute_string(
-            "\n".join(statements), cursor_class=VerboseCursor
+        if errors:
+            for error in errors:
+                logger.info("Statement compilation error: %s", error)
+                cli_console.warning(error)
+            raise ClickException("SQL rendering error")
+
+        is_single_statement = not (stmt_count > 1)
+        return is_single_statement, self._execute_string(
+            "\n".join(compiled_statements),
+            cursor_class=VerboseCursor,
         )