+++ tests_integration/sql/test_sql.py
@@ -140,15 +140,35 @@ def test_queries_are_streamed_to_output(
 
 
 @pytest.mark.integration
-def test_trailing_comments_queries(runner, snowflake_session, test_root_path):
-    trailin_comment_query = "select 1;\n\n-- trailing comment\n"
-    result = runner.invoke_with_connection_json(["sql", "-q", trailin_comment_query])
+@pytest.mark.parametrize(
+    "query, expected",
+    (
+        pytest.param(
+            "select 1; -- trailing comment\n",
+            [
+                {"1": 1},
+            ],
+            id="single query",
+        ),
+        pytest.param(
+            "select 1; --comment\n select 2; \n -- trailing comment\n",
+            [
+                [
+                    {"1": 1},
+                ],
+                [
+                    {"2": 2},
+                ],
+            ],
+        ),
+    ),
+)
+def test_trailing_comments_queries(runner, query, expected):
+    result = runner.invoke_with_connection_json(
+        ["sql", "-q", query, "--format", "JSON"]
+    )
     assert result.exit_code == 0
-    assert result.json == [
-        [
-            {"1": 1},
-        ],
-    ]
+    assert result.json == expected, result.json
 
 
 @pytest.mark.integration
@@ -199,3 +219,20 @@ def test_sql_with_variables_from_project(runner, project_directory, template):
         )
         assert result.exit_code == 0, result.output
         assert result.json == [{"VAR": "Knights of Nii"}]
+
+
+@pytest.mark.integration
+def test_sql_source_command_from_user_input(runner, tmp_path_factory, snapshot):
+    include_file = tmp_path_factory.mktemp("data") / "include.sql"
+    include_file.write_text("select 42;")
+
+    result = runner.invoke_with_connection(
+        (
+            "sql",
+            "-q",
+            f"select 1; !source {include_file.as_posix()}; select 3;",
+        )
+    )
+
+    assert result.output == snapshot
+    assert result.exit_code == 0, result.output